Reorganize! Also delete the pointless login page
